Fix EthPearl damage; add purchase individual cosmetics; misc cleanup

+++ core/src/main/java/be/isach/ultracosmetics/util/Particles.java
@@ -1137,7 +1137,7 @@ public enum Particles {
          */
         @Override
         public float getValueX() {
-            return (float) red / 255F;
+            return red / 255F;
         }
 
         /**
@@ -1147,7 +1147,7 @@ public enum Particles {
          */
         @Override
         public float getValueY() {
-            return (float) green / 255F;
+            return green / 255F;
         }
 
         /**
@@ -1157,7 +1157,7 @@ public enum Particles {
          */
         @Override
         public float getValueZ() {
-            return (float) blue / 255F;
+            return blue / 255F;
         }
     }
 
@@ -1195,7 +1195,7 @@ public enum Particles {
          */
         @Override
         public float getValueX() {
-            return (float) note / 24F;
+            return note / 24F;
         }
 
         /**
